requirements.txt
asgiref==3.7.2
Django==4.2.4
django-autoslug==1.9.9
django-etc==1.4.0
django-hitcount==1.3.5
django-modeltranslation==0.18.11
Pillow==10.0.0
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.7.1
whitenoise==6.5.0
